feat(compiler): `AutoBeInterfaceCompiler` with `prettier` formatting.
