+++ assets/documents/README.md
@@ -1,0 +1,66 @@
+# AutoBE Generated Backend Server
+
+![AutoBE Logo](https://github.com/user-attachments/assets/a90d14be-fd50-4dc7-ae9d-ca66c2124f31)
+
+A backend repository generated by [`@autobe`](https://github.com/wrtnlabs/autobe).
+
+This backend program was automatically generated using [`@autobe`](https://github.com/wrtnlabs/autobe), the AI vibe coding agent for backend servers of below stack.
+
+- TypeScript
+- NestJS / Nestia
+- Prisma
+- Postgres
+
+## Project Structure
+
+This template project has categorized directories like below.
+
+As you can see from the below, all of the Backend source files are placed into the [src](src/) directory. When you build the TypeScript source files, compiled files would be placed into the `lib` directory following the [tsconfig.json](tsconfig.json) configuration. Otherwise you build client [SDK](#32-sdk) library for npm publishing and their compiled files would be placed into the [packages](packages) directory.
+
+  - [packages/api/](packages/api): SDK module built by `npm run build:api`
+  - [src/](src): Backend source directory
+    - [src/api/](src/api/): Client SDK that would be published to the `@ORGANIZATION/PROJECT-api`
+      - [**src/api/functional/**](src/api/functional/): API functions generated by the [`nestia`](https://github.com/samchon/nestia)
+      - [**src/api/structures/**](src/api/structures/): DTO structures
+    - [src/controllers/](src/controllers/): Controller classes of the Main Program
+  - [**test/**](test): Test Automation Program
+    - [test/features](test/features): List of test functions
+  - [nestia.config.ts](nestia.config.ts): Configuration file of [`nestia`](https://github.com/samchon/nestia)
+  - [package.json](package.json): NPM configuration
+  - [tsconfig.json](tsconfig.json): TypeScript configuration for the main program
+
+## NPM Run Commands
+
+List of the run commands defined in the [package.json](package.json) are like below:
+
+  - Test
+    - **`test`**: Run test automation program
+    - `benchmark`: Run performance benchmark program
+  - Build
+    - `build`: Build everything
+    - `build:main`: Build main program (`src` directory)
+    - `build:test` Build test automation program (`test` directory)
+    - `build:sdk`: Build SDK into main program only
+    - `build:swagger`: Build Swagger Documents
+    - **`dev`**: Incremental build for development (test program)
+  - Deploy
+    - `package:api`: Build and deploy the SDK library to the NPM
+    - `start`: Start the backend server
+    - `start:dev`: Start the backend server with incremental build and reload
+  - Webpack
+    - `webpack`: Run webpack bundler
+    - `webpack:start`: Start the backend server built by webpack
+    - `webpack:test`: Run test program to the webpack built
+
+## Specialization
+
+Transform this template project to be yours.
+
+When you've created a new backend project through this template project, you can specialize it to be suitable for you by changing some words. Replace below words through IDE specific function like `Edit > Replace in Files` (*Ctrl + Shift + H*), who've been supported by the VSCode.
+
+| Before       | After
+|--------------|----------------------------------------
+| ORGANIZATION | Your account or corporation name
+| PROJECT      | Your own project name
+| AUTHOR       | Author name
+| https://github.com/samchon/nestia-start | Your repository URL
