+++ packages/agent/src/factory/index.ts
@@ -1,3 +1,2 @@
-export * from "./createOpenApiDocument";
 export * from "./createAgenticaHistory";
 export * from "./invertOpenApiDocument";
