+++ packages/compiler/src/AutoBeCompiler.ts
@@ -1,5 +1,7 @@
 import {
+  AutoBeOpenApi,
   IAutoBeCompiler,
+  IAutoBeInterfaceCompiler,
   IAutoBePrismaCompiler,
   IAutoBePrismaCompilerProps,
   IAutoBePrismaCompilerResult,
@@ -8,11 +10,14 @@ import {
   IAutoBeTypeScriptCompilerResult,
 } from "@autobe/interface";
 
+import { AutoBeInterfaceCompiler } from "./AutoBeInterfaceCompiler";
 import { AutoBePrismaCompiler } from "./AutoBePrismaCompiler";
 import { AutoBeTypeScriptCompiler } from "./AutoBeTypeScriptCompiler";
 
 export class AutoBeCompiler implements IAutoBeCompiler {
   private readonly prisma_: IAutoBePrismaCompiler = new AutoBePrismaCompiler();
+  private readonly interface_: IAutoBeInterfaceCompiler =
+    new AutoBeInterfaceCompiler();
   private readonly typescript_: IAutoBeTypeScriptCompiler =
     new AutoBeTypeScriptCompiler();
 
@@ -20,6 +25,12 @@ export class AutoBeCompiler implements IAutoBeCompiler {
     props: IAutoBePrismaCompilerProps,
   ): Promise<IAutoBePrismaCompilerResult> {
     return this.prisma_.compile(props);
+  }
+
+  public interface(
+    document: AutoBeOpenApi.IDocument,
+  ): Promise<Record<string, string>> {
+    return this.interface_.compile(document);
   }
 
   public typescript(
