+++ packages/compiler/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./AutoBeCompiler";
 export * from "./AutoBeTypeScriptCompiler";
+export * from "./AutoBeInterfaceCompiler";
 export * from "./AutoBePrismaCompiler";
