+++ packages/interface/src/compiler/IAutoBeCompiler.ts
@@ -1,3 +1,4 @@
+import { AutoBeOpenApi } from "../openapi";
 import { IAutoBePrismaCompilerProps } from "./IAutoBePrismaCompilerProps";
 import { IAutoBePrismaCompilerResult } from "./IAutoBePrismaCompilerResult";
 import { IAutoBeTypeScriptCompilerProps } from "./IAutoBeTypeScriptCompilerProps";
@@ -7,6 +8,9 @@ export interface IAutoBeCompiler {
   prisma(
     props: IAutoBePrismaCompilerProps,
   ): Promise<IAutoBePrismaCompilerResult>;
+
+  interface(document: AutoBeOpenApi.IDocument): Promise<Record<string, string>>;
+
   typescript(
     props: IAutoBeTypeScriptCompilerProps,
   ): Promise<IAutoBeTypeScriptCompilerResult>;
