+++ packages/interface/src/compiler/IAutoBeInterfaceCompiler.ts
@@ -1,0 +1,5 @@
+import { AutoBeOpenApi } from "../openapi";
+
+export interface IAutoBeInterfaceCompiler {
+  compile(document: AutoBeOpenApi.IDocument): Promise<Record<string, string>>;
+}
