+++ packages/interface/src/compiler/index.ts
@@ -2,6 +2,7 @@ export * from "./IAutoBeCompiler";
 export * from "./IAutoBeTypeScriptCompiler";
 export * from "./IAutoBeTypeScriptCompilerProps";
 export * from "./IAutoBeTypeScriptCompilerResult";
+export * from "./IAutoBeInterfaceCompiler";
 export * from "./IAutoBePrismaCompiler";
 export * from "./IAutoBePrismaCompilerProps";
 export * from "./IAutoBePrismaCompilerResult";
