+++ test/package.json
@@ -29,7 +29,6 @@
     "uuid": "^11.1.0"
   },
   "devDependencies": {
-    "@nestia/migrate": "^6.0.1",
     "@types/node": "^22.15.17",
     "@types/uuid": "^10.0.0",
     "rimraf": "^6.0.1",
